Split text block editor components in separate modules
